requirements.txt
pydantic>=2,<3
python-docx>=0.8.11
pytest>=8
